+++ Sources/StreamFeeds/FeedsClient.swift
@@ -39,6 +39,7 @@ public final class FeedsClient: Sendable {
     let feedsRepository: FeedsRepository
     let moderationRepository: ModerationRepository
     let pollsRepository: PollsRepository
+    let ownCapabilitiesRepository: OwnCapabilitiesRepository
 
     private let _token: AllocatedUnfairLock<UserToken>
     private let _userAuth = AllocatedUnfairLock<UserAuth?>(nil)
@@ -127,14 +128,24 @@ public final class FeedsClient: Sendable {
         commentsRepository = CommentsRepository(apiClient: apiClient)
         devicesRepository = DevicesRepository(devicesClient: devicesClient)
         feedsRepository = FeedsRepository(apiClient: apiClient)
-        pollsRepository = PollsRepository(apiClient: apiClient)
         moderationRepository = ModerationRepository(apiClient: apiClient)
+        ownCapabilitiesRepository = OwnCapabilitiesRepository(apiClient: apiClient)
+        pollsRepository = PollsRepository(apiClient: apiClient)
         
         moderation = Moderation(apiClient: apiClient)
         
         eventsMiddleware.add(subscriber: self)
         eventsMiddleware.add(subscriber: stateLayerEventPublisher)
         eventNotificationCenter.add(middlewares: [eventsMiddleware])
+        
+        stateLayerEventPublisher.addMiddlewares(
+            [
+                OwnCapabilitiesStateLayerEventMiddleware(
+                    ownCapabilitiesRepository: ownCapabilitiesRepository,
+                    sendEvent: { [weak stateLayerEventPublisher] in await stateLayerEventPublisher?.sendEvent($0) }
+                )
+            ]
+        )
     }
     
     // MARK: - Connecting the User
@@ -374,14 +385,20 @@ public final class FeedsClient: Sendable {
         ActivityReactionList(query: query, client: self)
     }
     
-    /// Adds a new activity to the specified feeds.
+    // MARK: - Activity Batch Operations
+    
+    /// Adds a new activity to one or multiple feeds.
     ///
-    /// - Parameter request: The request containing the activity data to add
+    /// - Parameter request: The request containing the activity data to add and the list of feeds
     /// - Returns: A response containing the created activity
     /// - Throws: `APIError` if the network request fails or the server returns an error
     @discardableResult
-    public func addActivity(request: AddActivityRequest) async throws -> AddActivityResponse {
-        try await apiClient.addActivity(addActivityRequest: request)
+    public func addActivity(request: AddActivityRequest) async throws -> ActivityData {
+        let activityData = try await activitiesRepository.addActivity(request: request)
+        for feed in request.feeds.map(FeedId.init) {
+            await stateLayerEventPublisher.sendEvent(.activityAdded(activityData, feed))
+        }
+        return activityData
     }
     
     /// Upserts (inserts or updates) multiple activities.
@@ -391,7 +408,16 @@ public final class FeedsClient: Sendable {
     /// - Throws: `APIError` if the network request fails or the server returns an error
     @discardableResult
     public func upsertActivities(_ activities: [ActivityRequest]) async throws -> [ActivityData] {
-        try await activitiesRepository.upsertActivities(activities)
+        let activities = try await activitiesRepository.upsertActivities(activities)
+        let updates = activities.reduce(into: ModelUpdates<ActivityData>()) { partialResult, activityData in
+            if activityData.updatedAt.timeIntervalSince(activityData.createdAt) > 0.1 || activityData.editedAt != nil {
+                partialResult.updated.append(activityData)
+            } else {
+                partialResult.added.append(activityData)
+            }
+        }
+        await stateLayerEventPublisher.sendEvent(.activityBatchUpdate(updates))
+        return activities
     }
     
     /// Deletes multiple activities from the specified feeds.
@@ -400,8 +426,10 @@ public final class FeedsClient: Sendable {
     /// - Returns: A response confirming the deletion of activities
     /// - Throws: `APIError` if the network request fails or the server returns an error
     @discardableResult
-    public func deleteActivities(request: DeleteActivitiesRequest) async throws -> DeleteActivitiesResponse {
-        try await apiClient.deleteActivities(deleteActivitiesRequest: request)
+    public func deleteActivities(request: DeleteActivitiesRequest) async throws -> Set<String> {
+        let response = try await apiClient.deleteActivities(deleteActivitiesRequest: request)
+        await stateLayerEventPublisher.sendEvent(.activityBatchUpdate(.init(added: [], removedIds: response.deletedIds, updated: [])))
+        return Set(response.deletedIds)
     }
     
     // MARK: - Bookmark Lists