+++ Tests/StreamFeedsTests/Docs/Snippets-03-01-Activities.swift
@@ -3,6 +3,7 @@
 //
 
 import Combine
+import Foundation
 import StreamCore
 import StreamFeeds